requirements.txt
streamlit
playwright
google-generativeai
python-dotenv
requests
beautifulsoup4
pyperclip
Pillow
pyinstaller
